requirements.txt
torch
scipy
onnxruntime
joblib
pyyaml
easydict
python-box
tqdm
pygame
zmq
pynput
pydantic
mujoco
msgpack
msgpack-numpy
colorlog
imageio
